G.5.1: Fix stub headers and build libmako_core successfully
Stub header fixes:
- sstream: Remove duplicate manipulator definitions (now in iostream)
- stdint.h: Fix int64_t/uint64_t to use long on x86_64 (matching cstdint)
- stdlib.h: Guard rand_r to avoid conflict with cstdlib

CLI improvements:
- Add stubs directory as system include for proper C++ parsing
- Add debug mode (FRAGILE_DEBUG=1) for compiler command logging

Build configuration:
- Fix Mako fragile.toml CONFIG_H path (mako/config/config-perf.h)
- Remove src/mako/lib from libmako_core includes to avoid assert.h
  shadowing the system header

Successfully builds:
- librrr: 19 object files, 11MB static library
- libmako_core: 12 object files, 4.7MB static library

+++ crates/fragile-clang/stubs/stdint.h
@@ -5,15 +5,16 @@
 #define _STDINT_H
 
 // Exact width integer types
+// Note: On x86_64 Linux, long is 64-bit, so int64_t should be long, not long long
 typedef signed char int8_t;
 typedef short int16_t;
 typedef int int32_t;
-typedef long long int64_t;
+typedef long int64_t;  // long is 64-bit on x86_64
 
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long long uint64_t;
+typedef unsigned long uint64_t;  // unsigned long is 64-bit on x86_64
 
 // Minimum-width integer types
 typedef int8_t int_least8_t;
@@ -42,8 +43,9 @@ typedef long intptr_t;
 typedef unsigned long uintptr_t;
 
 // Greatest-width integer types
-typedef long long intmax_t;
-typedef unsigned long long uintmax_t;
+// Note: On x86_64 Linux, these should be long to match cstdint
+typedef long intmax_t;
+typedef unsigned long uintmax_t;
 
 // Limits of exact-width integer types
 #define INT8_MIN (-128)
